+++ example/index.php
@@ -1,0 +1,28 @@
+<?php
+
+require __DIR__ . '/bootstrap.php';
+
+// do stuff?
+
+?>
+<html>
+    <head>
+        <meta charset="UTF-8" />
+        <title>MedoolsRouter Example</title>
+    </head>
+    <body>
+        <h1>Hello World!</h1>
+        <p>
+            See the documentation <a href="https://github.com/aryelgois/medools-router">here</a>.
+        </p>
+        <p>
+            See the <a href="README.md">README.md</a>.
+        </p>
+        <p>
+            Access the Router <a href="api/">here</a>.
+        </p>
+
+        <!-- This script would load data from /api to modify the page -->
+        <!-- <script src="/main.js"></script> -->
+    </body>
+</html>