+++ CHANGELOG.md
@@ -7,3 +7,36 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+## [0.1.0] - 2018-03-31
+
+### Added
+- Classes:
+  - Authentication
+  - Authorization
+  - Controller
+  - Router
+  - RouterException
+  - Resource
+  - Response
+- Databases:
+  - authentication
+- Dependencies:
+  - [aryelgois/yasql-php]
+  - [firebase/php-jwt]
+- Documentation
+- Example
+- [Medools][aryelgois/medools] config file
+- Platform requirement: `ext-zlib`
+
+### Changed
+- Update dependencies
+- Namespace to `aryelgois\MedoolsRouter`
+
+
+[Unreleased]: https://github.com/aryelgois/medools-router/compare/v0.1.0...develop
+[0.1.0]: https://github.com/aryelgois/medools-router/compare/d281bb5dbc8c58b28db680b3700664217a88eb6d...v0.1.0
+
+[aryelgois/medools]: https://github.com/aryelgois/Medools
+[aryelgois/yasql-php]: https://github.com/aryelgois/yasql-php
+[firebase/php-jwt]: https://github.com/firebase/php-jwt